requirements.txt
pytest==7.4.3
pytest-html==4.1.1
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
requests==2.31.0
selenium==4.16.0
allure-pytest==2.13.2
faker==21.0.0
pyyaml==6.0.1
python-dotenv==1.0.0
